--- a/quarkus/config-api/src/main/java/org/keycloak/config/database/Database.java
+++ b/quarkus/config-api/src/main/java/org/keycloak/config/database/Database.java
@@ -18,6 +18,7 @@
 package org.keycloak.config.database;
 
 import java.io.File;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
@@ -49,6 +50,12 @@ public final class Database {
         }
 
         return false;
+    }
+
+    public static Optional<Vendor> getVendorByDbKind(String dbKind) {
+        return Arrays.stream(Vendor.values())
+                .filter(v -> v.isOfKind(dbKind))
+                .findAny();
     }
 
     public static Optional<String> getDatabaseKind(String alias) {