--- a/quarkus/runtime/src/test/java/org/keycloak/quarkus/runtime/configuration/test/IgnoredArtifactsTest.java
+++ b/quarkus/runtime/src/test/java/org/keycloak/quarkus/runtime/configuration/test/IgnoredArtifactsTest.java
@@ -17,15 +17,27 @@
 
 package org.keycloak.quarkus.runtime.configuration.test;
 
+import org.hamcrest.CoreMatchers;
 import org.junit.Test;
 import org.keycloak.common.Profile;
 import org.keycloak.common.profile.PropertiesProfileConfigResolver;
+import org.keycloak.config.DatabaseOptions;
 import org.keycloak.quarkus.runtime.configuration.IgnoredArtifacts;
+import org.keycloak.quarkus.runtime.configuration.MicroProfileConfigProvider;
 
+import java.util.HashSet;
 import java.util.Properties;
+import java.util.Set;
 
 import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.keycloak.quarkus.runtime.configuration.IgnoredArtifacts.JDBC_H2;
+import static org.keycloak.quarkus.runtime.configuration.IgnoredArtifacts.JDBC_MARIADB;
+import static org.keycloak.quarkus.runtime.configuration.IgnoredArtifacts.JDBC_MSSQL;
+import static org.keycloak.quarkus.runtime.configuration.IgnoredArtifacts.JDBC_MYSQL;
+import static org.keycloak.quarkus.runtime.configuration.IgnoredArtifacts.JDBC_ORACLE;
+import static org.keycloak.quarkus.runtime.configuration.IgnoredArtifacts.JDBC_POSTGRES;
 
 public class IgnoredArtifactsTest {
 
@@ -48,5 +60,53 @@ public class IgnoredArtifactsTest {
 
         var ignoredArtifacts = IgnoredArtifacts.getDefaultIgnoredArtifacts();
         assertThat(ignoredArtifacts.containsAll(IgnoredArtifacts.FIPS_ENABLED), is(true));
+    }
+
+    @Test
+    public void jdbcH2() {
+        assertJdbc("h2", JDBC_H2);
+    }
+
+    @Test
+    public void jdbcMssql() {
+        assertJdbc("mssql", JDBC_MSSQL);
+    }
+
+    @Test
+    public void jdbcMariadb() {
+        assertJdbc("mariadb", JDBC_MARIADB);
+    }
+
+    @Test
+    public void jdbcMysql() {
+        assertJdbc("mysql", JDBC_MYSQL);
+    }
+
+    @Test
+    public void jdbcOracle() {
+        assertJdbc("oracle", JDBC_ORACLE);
+    }
+
+    @Test
+    public void jdbcPostgres() {
+        assertJdbc("postgres", JDBC_POSTGRES);
+    }
+
+    private void assertJdbc(String vendor, Set<String> notIgnored) {
+        System.setProperty(MicroProfileConfigProvider.NS_KEYCLOAK_PREFIX + DatabaseOptions.DB.getKey(), vendor);
+        try {
+            final var resultArtifacts = IgnoredArtifacts.getDefaultIgnoredArtifacts();
+            assertThat(String.format("Ignored artifacts does not comply with the specified artifacts for '%s' JDBC driver", vendor),
+                    resultArtifacts,
+                    not(CoreMatchers.hasItems(notIgnored.toArray(new String[0]))));
+
+            final var includedArtifacts = new HashSet<>(IgnoredArtifacts.JDBC_DRIVERS);
+            includedArtifacts.removeAll(notIgnored);
+            assertThat("Ignored artifacts does not contain items for the other JDBC drivers",
+                    resultArtifacts,
+                    CoreMatchers.hasItems(includedArtifacts.toArray(new String[0])));
+        } finally {
+            System.setProperty(MicroProfileConfigProvider.NS_KEYCLOAK_PREFIX + DatabaseOptions.DB.getKey(), "");
+        }
     }
 }