Closes #22442