@mabartos 2023-09-14
@ahus1 Thanks for your points. There are performance benefits, indeed. I'll describe it more tomorrow.