@mabartos 2023-09-15
This change brings quite a lot of benefits to us. First of all, the execution of build, static-init, and runtime build steps is done only for the specified DB. By this change, Quarkus is not aware of any other JDBC extension except the one for the chosen DB, so there's no need to execute these build steps, which could produce additional beans, service providers, configuration handling, agroal connection settings, and much more.

To summarize, we can prevent the execution of unused build steps, create additional resources, reduce memory allocations, GC executions, C1/C2 executions... 

I've made a brief analysis of the performance (few executions followed by a comparison):

**Startup**

- The significant improvement in startup is for AgroalProcessor as we saved ~80ms (cnt of CPU samples decreased by ~66%). 
- The overall improvement is ~100ms. 

**Heap** 
- I've encountered a small peek during the startup
- I assume objects provided during the build steps are destroyed after a few GC executions. 
- There's no peek anymore

Old
<img src="https://github.com/keycloak/keycloak/assets/38039883/3cf80206-c6a9-4aa2-b4f0-4728c034d801"/>

New
<img src="https://github.com/keycloak/keycloak/assets/38039883/4dbb318c-9d7e-4951-a58d-19e67bd5be75"/>

**GC** - GC executions reduced about ~20%.
**Compiler** - compilations reduced about ~13%

**IMPORTANT**: However, I expected significant changes around the class loading with the presence of the JDBC drivers' JARs. I think specifying the JDBC drivers into the ignored artifacts does not take any effect as it's probably used only for the custom Quarkus ClassLoaders anyway, which doesn't bring them into consideration if they do not have to be used. 

IMO, they can be removed from the codebase and reduce the need for manageability of it; I'll verify it.

@ahus1 
> Will this prevent some of the warning messages we've seen earlier when a DB driver is removed?

Yes, and much more as the build steps for different JDBC extensions are not executed at all.

> I could see the need for some users to change from an included DB driver to a self-provided DB driver. Would this change enable this, or is it unrelated?

I think it's unrelated to this PR as the logic behind using the custom JDBC drivers has not changed.