@mabartos 2023-09-18
@vmuzikar @ahus1 I've put some details in the comment above.

This PR is in a final state, and the JDBC Drivers' JARs have to be specified there.
IMO, it is definitely worth putting these changes into Keycloak despite the manageability it requires. 

@vmuzikar @ahus1 WDYT?