@ahus1 2023-09-18
@mabartos - looking at the images with with "new" using more memory than "old", it seems they have been swapped? 

With 100 ms improvement in built time, when we ask performance interested users to use pre-built images where possible, and ~7 MB of RAM during build time (which is not relevant at run time), I'm still not convinced. 

At the same time, I won't veto it if the CND still wants to merge it.