@ahus1 2023-09-18
@mabartos - thank you for the additional explanations. No need for meeting, let's wait for @vmuzikar's response only.