@mabartos 2023-09-20
Okay, it seems I haven't explained it accurately, as it may cause some confusion around it; my bad. 

I'll answer the questions here and probably summarize the provided changes and benefits in a more clear way. Probably even start the Keycloak Benchmark, if it provides the functionality to compare distros with/without these changes (like startup time, memory footprint at the start, GC activities,...?)

> The ignored artifacts should really affect only build time. They are in the end mapped to [quarkus.class-loading.removed-artifacts](https://quarkus.io/guides/class-loading-reference#quarkus-class-loading-configuration-class-loading-config_quarkus.class-loading.removed-artifacts) which is build time option.

@vmuzikar
Even when it's a build time option, it doesn't mean it doesn't affect runtime, of course. Quite the opposite; that's why we can execute the augmented application. The ignored artifacts are banned for all Quarkus ClassLoaders, so the Base Runtime and Runtime ClassLoader as well. It means they're completely removed from the application execution. To be more specific, no beans are created in static-init/runtime-init bootstrap phase defined by these extensions, ServiceLoaders do not manage any provider included in these JARs,...

> If I understand correctly, you were testing runtime with auto-build enabled. Is that correct? I'd like to see if there's any performance improvement in case of optimized image, i.e. if we omit the build phase.

Not at all. There wasn't any build phase done during the testing; the optimized version was used.

> What slightly concerns me is that it appears after this change more memory is allocated (at least initially until things settle down)? I know it removed the initial spike but the overall memory is higher?

Nope, the overall memory was quite the same after some time. These pictures were only for demonstrating the initial peak. Described in comment https://github.com/keycloak/keycloak/pull/22443#issuecomment-1723624265.