@mabartos 2023-09-20
This PR is starting to be quite messy as there are certain concerns about perf improvements it might bring vs. manageability.

I'll try to find out better approach to gathering performance results to emphasize it more :stuck_out_tongue_closed_eyes: (will try to leverage Keycloak Benchmark)