@abstractj 2023-10-10
@mabartos based on the discussions here I believe we have a couple of action items to follow up from my perspective (maybe others will disagree):

1. Address flaky tests: investigate the unreported flaky tests mentioned in the pull request to determine if they are affected by the changes proposed.
2. Clarity about the performance benefits: provide a more detailed analysis of the performance improvements, highlighting the specific improvements in startup time, memory allocations, GC executions...
3. Run tests with auto-build enabled to determine if there are any performance improvements in the case of an optimized image.

@vmuzikar do the action items above make any sense?