@mabartos 2023-11-02
# Performance analysis
This performance analysis touches on multiple aspects of the application's performance. JFR (Java Flight Recorder), Async profiler, and Podman stats, for testing performance for optimized Keycloak image, were the main tools used for the performance analysis. 

The expected improvement is related to the ignoring JARs for JDBC drivers and its related Quarkus extensions. We will prevent additional Class loading, C1/C2 executions, GC cycles, and memory allocations.

TERMINOLOGY: In this analysis, I'll compare behavior before and after these changes. They will be marked as **OLD** and **NEW**.

## Profiling
In this section, I'll describe the highlights of benefits these changes bring to us analyzed with the usage of JFR and Async profiler. 

### CPU
I had the possibility to test it on two different machines as we can also see the difference between more and less performant agents. In the tables below, there are described main differences in used CPU samples during the profiling. CPU samples basically represent a count of encountered snapshots in time, where we can find out what operations' execution occupies the CPU the most.

CPU sampling 1ms/sample. The numbers in the table below represent a count of CPU samples. 

**Lenovo T490s (Intel Core i7 Pro 8xCPU, 16GB Mem)**
<table>
    <tr>
       <td align="center">Method</td>
       <td align="center">OLD</td>
       <td align="center">NEW</td>
       <td align="center">DIFF</td>
    </tr>
    <tr>
      <td>AgroalProcessor.generateDataSourceBeans</td>
      <td>163</td>
      <td>34</td>
      <td>-129</td>
    </tr>
    <tr>
      <td>  --- (inner) DataSources.loadDriversInTCCL</td>
      <td>125</td>
      <td>2</td>
      <td>-123</td>
    </tr>
     <tr>
      <td>HibernateOrmRecorder.created</td>
      <td>748</td>
      <td>712</td>
      <td>-36</td>
    </tr>
    <tr>
      <td>DataSource.init</td>
      <td>15</td>
      <td>11</td>
      <td>-4</td>
    </tr>
  </table>
  
**Lenovo X1 Nano (12th gen Intel Core i7-1280P 20xCPU, 32GB Mem)**
<table>
    <tr>
       <td align="center">Method</td>
       <td align="center">OLD</td>
       <td align="center">NEW</td>
       <td align="center">DIFF</td>
    </tr>
    <tr>
      <td>AgroalProcessor.generateDataSourceBeans</td>
      <td>70</td>
      <td>20</td>
      <td>-50</td>
    </tr>
    <tr>
      <td>  --- (inner) DataSources.loadDriversInTCCL</td>
      <td>53</td>
      <td>1</td>
      <td>-52</td>
    </tr>
     <tr>
      <td>HibernateOrmRecorder.created</td>
      <td>388</td>
      <td>345</td>
      <td>-43</td>
    </tr>
    <tr>
      <td>DataSource.init</td>
      <td>15</td>
      <td>11</td>
      <td>-4</td>
    </tr>
  </table>  
  
### Memory allocations
The biggest difference in memory allocations is in the `AgroalProcessor` as well. As inside of the method, the call for `DataSources.loadDriversInTCCL()` is done and it loads all JDBC drivers found by ServiceLoader. As we will load only the specific JDBC driver we want, the memory allocation for it is drastically decreased.

<table>
    <tr>
       <td align="center">Method</td>
       <td align="center">OLD</td>
       <td align="center">NEW</td>
       <td align="center">DIFF</td>
    </tr>
    <tr>
      <td>AgroalProcessor.generateDataSourceBeans</td>
      <td>1.7 MB</td>
      <td>9.5 MB</td>
      <td>-7.8 MB</td>
    </tr>
  </table>
  
### Startup time
The biggest improvement in the startup time is also related to the AgroalProcessor, as the CPU samples decreased.  
 
 <table>
    <tr>
       <td align="center">Build item</td>
       <td align="center">OLD</td>
       <td align="center">NEW</td>
       <td align="center">DIFF</td>
    </tr>
    <tr>
      <td>AgroalProcessor.generateDataSourceBeans</td>
      <td>124 ms</td>
      <td>23 ms</td>
      <td>-101 ms</td>
    </tr>
    <tr>
      <td>ArcProcessor.generateResources</td>
      <td>456 ms</td>
      <td>414 ms</td>
      <td>-42 ms</td>
    </tr>
  </table>
  
  NOTE: Only approximately; it differs execution by execution, but most of the time very similar. 
  
  ### Other improvements
I've spotted also other improvements in different areas of the application and I can share some of them here. I haven't encountered any negative performance impact provided by these changes.

 <table>
    <tr>
       <td align="center">Aspects</td>
       <td align="center">OLD</td>
       <td align="center">NEW</td>
       <td align="center">DIFF</td>
    </tr>
    <tr>
      <td>CPU Load Max</td>
      <td>43.4 %</td>
      <td>40.2 %</td>
      <td>-3.2 %</td>
    </tr>
    <tr>
      <td>Loaded class count</td>
      <td>19049</td>
      <td>18763</td>
      <td>-286</td>
    </tr>
    <tr>
      <td>Quarkus Class Loader Chunk size</td>
      <td>10.4 MiB</td>
      <td>10.3 MiB</td>
      <td>-0.1 MiB</td>
    </tr>
    <tr>
      <td>Garbage Collection count</td>
      <td>26</td>
      <td>25</td>
      <td>-1</td>
    </tr>
    <tr>
      <td>Garbage Collection duration (ParallelGC)</td>
      <td>39.089 ms</td>
      <td>33.197 ms</td>
      <td>-5.89 ms</td>
    </tr>
  </table>
  
## Optimized image
In this section, I'll describe the highlights of benefits these changes bring to us analyzed by Podman statistics (`podman stats`) for the optimized Keycloak image. The build phase was executed in the _builder_ phase of a Dockerfile and tested was only the runtime behavior for the optimized image with H2 in-memory DB without executing any DB schema init or upgrade. 

#### Benchmark settings

- Analyzing CPU utilization and Memory usage
- Snapshots taken every 1s (`podman stats --no-reset -i 1`)
- Optimized Keycloak images
- **OLD** (Keycloak main) and **NEW** (this PR) image was executed 3x 
- Computed average of these values for every snapshot

#### Memory usage
You can noticed the memory usage is lower for the NEW image and after ~10s, it's more or less stabilized as the old one. 
We don't encounter the initial peak for the memory usage due to the JDBC Driver JARs inits. No comprehensive GC cycles.
The memory usage is related to the whole image; with JVM. 

<img width="450" src="https://github.com/keycloak/keycloak/assets/38039883/c1256c6f-e781-46ed-a728-1d8809f1a805"/>

#### Average CPU Usage
The average CPU is the same in ~0-2s, but then, I assume, the JDBC Drivers loading with followed GC execution is done and the average CPU usage is higher for the OLD image. For more info how the Average CPU is computed for the Podman/Docker stats may be found here[1]. 

<img width="450" src="https://github.com/keycloak/keycloak/assets/38039883/3bcc2464-cec2-4a73-a021-1b1e5b57c793"/>

[1] https://stackoverflow.com/questions/47401648/docker-stats-with-cpu-percentage-more-than-100

## Testing device info
```
CPU(s):                  20
  On-line CPU(s) list:   0-19
Vendor ID:               GenuineIntel
  Model name:            12th Gen Intel(R) Core(TM) i7-1280P
    CPU family:          6
    Model:               154
    Thread(s) per core:  2
    Core(s) per socket:  14
    Socket(s):           1
    Stepping:            3
    CPU(s) scaling MHz:  31%
    CPU max MHz:         4800.0000
    CPU min MHz:         400.0000
Memory:                  32GB    
```

## Summary 

@vmuzikar @ahus1 @abstractj
Here is the detailed performance analysis for the benefits it will brings to us. I hope it's sufficient to show that it'll improve the overall performance. Please let me know, if you're more curious about it, you want the testing data or you have some input on this :))

I'd be glad to answer your questions and react to your comments. 

I think it worths to include it into Keycloak even there will be slight need for maintanance. WDYT?